requirements.txt
astroid==1.6.5
Django==2.0.6
django-extensions==2.0.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
pkg-resources==0.0.0
pylint==1.9.2
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
wrapt==1.10.11
